Move key error code to submodule
